import: fixed review comments
